add missing/new ingredient
Decided to finish or further complete project while continuing to use userDefaults (for now). While practicing with realm. As I realised after working on realm for the last day and a half that it will require me to essentially start fresh and re-think my design. I'd like to be working on small improvements and making the app look nice but instead I feel like I've taken 10 steps back while working on realm. unsure how to proceed